requirements/dev.txt
# 개발용 라이브러리 목록
-r common.txt

django-debug-toolbar==4.3.0
ipython
django-environ==0.11.2